Add Insurance module
Add Insurance module.

This module does not need an API endpoint. This model does not save any requests, but is needed for defining the structure of what data can be received for a specific insurance.

Follow following JSON format to have a valid insurance model, where field_detail and its arguments are optional:
{
  "fields": [ 
    {
      "field_name": "kfz_bm",
      "field_title": "BM-Stufe",
      "field_type": "integer",
      "field_detail": {
        "min": 1,
        "max": 9
      }
    }
  ]
}