Add fields to User model
Add several fields to the User model:
- first_name
- last_name
- phone
- verified

The return string for an user is changed. An user now returns 'Firstname Lastname', or, if those variables are not set, it returns the username.

Phone number validation is added to the user module: check_phone_number().
For this, the module phonenumbers is added to the base requirements and to the django settings.

The new User model fields are added to the UserAdmin.