Add advisor field to user model
Add a field to the user model which is linking to itself for storing a customer advisor.

The field can also be blank/null and is set to null when a staff member is deleted.

For this field, every user can be selected and saved on every user.
For optimal implementation, it should be considered using a proxy model only containing users with 'utype' set to 7, which are staff members.
Due to only software administrators being able to manually change the value of an advisor field, this is not critical.