Add ChangeUserSerializer user data validation
Add custom validation of user data to the ChangeUserSerializer.

The data is validated using user.create_or_login.validated_user_data.

The serializer defines an update function, which uses the serializer's fields in order to update the user instance.

user.create_or_login.validated_user_data now skips the validation whether a field is required or not when stated that a user instance is only changed.