Add policy document field to submission model
Add a field for storing a policy's document file to the InsuranceSubmission model.

The field is called policy_document.
The create_path functionality is copied from the user model's profile picture field and should be generalized in the future (DRY).

The API at insurancesubmission.api.dev now includes the ChangeInsuranceSubmissionDetails view, which is accessible by administrative users and allows changing the active and denied boolean on a corresponding object, as well as setting/changing the policy id and the policy document.

The API endpoint is reached under insurancesubmission.api at <int:pk>/change/.
This is the first API endpoint, which takes the object identifier before the action definition. In the future, all API endpoints should be considered to be of this format.

The insurancesubmission admin is added the denied field to displayed and filterable fields, as well as the id field to searchable and order-able fields.

The sequence of the imports at the user model file is changed to be alphabetically ordered.