Add options and remove required at InsuranceSubmission Serializer
Add the option for adding an 'options' parameter to a insurance field, which accepts a list with all possible values, strings as well as numbers.
If the submitted content is not an option for this field, an error is returned to the client.

Change the handling of 'required' field parameter at the InsuranceSubmission Serializer.
The 'required' tag now defaults to True, and can be set to False if needed.
The required parameter should not be used at the moment, as the calculation needs every field.