Revert Changes on SubmitInsurance
Revert the changes on the SubmitInsurance endpoint.

Through this endpoint, only the customer can request a new insurance for himself. The customer should not be able to upload signed documents at this point, nor should he be able to submit templates.

Also the authentication data is missing in the new response when using the endpoint for user registration.

The code was made much cleaner through commit #7fdcaf67c95d00e8b331c3554dfd3038bb8d07e7 though, that was nice. Unfortunately, time's too expensive atm to change a running system.