Add endpoints for document sign process request
Add two endpoints whose purpose is to request the process of document signing:
- At /sign/submission/<int:pk>/request/, the signing process for all documents of a specific submission is initiated.
- At /sign/document/<int:pk>/request/, the signing process for all documents of a specific submission, starting with given document, is initiated.

The endpoints are not final. Functionality and data is missing, especially at RequestSignDocument.

Furthermore, an error at Document creation is fixed:
- Document.create() -> Document.objects.create()

A variable is changed which is returned after creating a document:
- policy id -> submission pk

As well as a typo is fixed:
- get_submission submissions -> submission