Finalize Payment PDF and document signing
Finalize the generation of the Payment PDF file and the document signing system.

I made use of two different PDF libraries:
- FPDF for generating a new PDF file containing necessary custom data, and
- PyPDF2 for combing the template file and the custom-generated file.

Both libraries are coming to use at Payment PDF as well as at PDF signing.

A template for Wiener Städtische Versicherungen is added. This could be improved - it currently doesn't contain text but only an image of the original document, which is encrypted and not editable.

A page_index value is added to the document model. This field identifies on what page of the document the signature must be placed, starting with index 0 for the first page.

The PDF file containing the signature only without any content is saved as <FileName>.pdf.signature.