Adaptions, workflow improvements and bug fixes
Add two fields to the insurance calculation for WSE:
- geldwerte, which adds 19€ on selection, and
- wertgegenstaende, which adds 75€ on selection.

Remove token from ProgressReport view, as the user must be authenticated anyway.

Implement signing a token on document signing, as this was missing before.

Fix PDF signing signature insertion bug, where only the signed page was added as the final document. Now, all pages of the document are added to the final document.

Start altering the "Beratungsprotokoll" PDF file, which is not finished yet. -> TODO