Add several changes
- Add Twilio implementation:
A new phone number can now be added through api/dev/user/phone/add/, and has to be verified using a code at endpoint api/dev/user/phone/verify/.
Twilio environment variables are added to the .env.example file and to the settings file.
The twilio package in current version is added to the base requirements.
The phone number cannot be changed on UserDetail anymore.

- Make the static files accessible to the outside
The two insurance logos are added for being served to the front-end on insurance submission query.

- Payment information is added to the insurance submission object, using a trimmed IBAN (last 4 digits only).

- An insurance submission can successfully be set to status '2' and '3'.

- A bug concerning document tokens is partially fixed, where a user could have multiple tokens due to creating tokens for different documents at the same time. This could be done more clean in fact.

- Integrate the provider id into the document creation processes.

- Add a signature at the last page of the Beratungsprotokoll.

- Change the font family of the payment document to Courier.